fix: decode money fields as float64 (plan_price, price_per_month)
The SpaceWeb API returns fractional prices (e.g. 0.9). VPS.PlanPrice and
VPSPlan.PricePerMonth were typed `int`, so a real List/refresh failed with
`cannot unmarshal number 0.9 into Go struct field .plan_price of type int`,
blocking terraform refresh of an existing node. Money is naturally
fractional → both fields are now float64. Fixtures made fractional (0.9 /
500.5) to lock the regression; List/AvailableConfig tests assert the values.

Contract note (pre-1.0): a widening of two response fields int→float64.

+++ config.go
@@ -17,16 +17,16 @@ type AvailableConfig struct {
 // VPSPlan is a purchasable VPS plan. Note SpaceWeb returns several numeric-ish
 // fields (cpu_cores, ram, volume_disk) as strings.
 type VPSPlan struct {
-	ID            int    `json:"id"`
-	Name          string `json:"name"`
-	PricePerMonth int    `json:"price_per_month"`
-	Category      string `json:"category"`
-	CPUCores      string `json:"cpu_cores"`
-	RAM           string `json:"ram"`
-	DiskType      string `json:"disk_type"`
-	VolumeDisk    string `json:"volume_disk"`
-	Datacenters   []int  `json:"datacenters"`
-	SoldOut       bool   `json:"sold_out"`
+	ID            int     `json:"id"`
+	Name          string  `json:"name"`
+	PricePerMonth float64 `json:"price_per_month"` // money: API returns fractional prices
+	Category      string  `json:"category"`
+	CPUCores      string  `json:"cpu_cores"`
+	RAM           string  `json:"ram"`
+	DiskType      string  `json:"disk_type"`
+	VolumeDisk    string  `json:"volume_disk"`
+	Datacenters   []int   `json:"datacenters"`
+	SoldOut       bool    `json:"sold_out"`
 }
 
 // OSOption is a selectable OS image.
